blockmodel functions add api, render and sw js-files, add height to reviewers card
