+++ src/css/reviews.css
@@ -53,7 +53,7 @@ color: #FAFAFA;
   border-color: #E0E0E0;
   padding: 32px 24px;
   width: 332px;
-  min-height: 302px;
+  height: 352px;
   box-sizing: border-box;
 }
